# cenagem-backend/openapi/cenagem-api.yaml
openapi: 3.1.0
info:
  title: CENAGEM API
  description: API oficial para historias clínicas familiares, agenda, estudios y adjuntos.
  version: 1.0.0
servers:
  - url: https://api.cenagem.gob.ar/api/v1
    description: Producción
  - url: http://localhost:3000/api/v1
    description: Desarrollo local
paths:
...
  /families/{familyId}/appointments:
    get:
      tags: [appointments]
      summary: Turnos de una familia
      parameters:
        - $ref: "#/components/parameters/FamilyId"
      responses:
        '200':
          description: Turnos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentListResponse"
    post:
      tags: [appointments]
      summary: Crear turno para familia
      parameters:
        - $ref: "#/components/parameters/FamilyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentCreateInput"
      responses:
        '201':
          description: Turno creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
...
  /families/{familyId}/studies:
    get:
      tags: [studies]
      summary: Estudios de una familia
      parameters:
        - $ref: "#/components/parameters/FamilyId"
      responses:
        '200':
          description: Estudios
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyListResponse"
    post:
      tags: [studies]
      summary: Registrar estudio
      parameters:
        - $ref: "#/components/parameters/FamilyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyCreateInput"
      responses:
        '201':
          description: Estudio creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Study"
  /studies/{studyId}:
    patch:
      tags: [studies]
      summary: Actualizar estudio
      parameters:
        - $ref: "#/components/parameters/StudyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyUpdateInput"
      responses:
        '200':
          description: Estudio actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Study"
    delete:
      tags: [studies]
      summary: Eliminar estudio
      parameters:
        - $ref: "#/components/parameters/StudyId"
      responses:
        '200':
          description: Estudio eliminado
...
components:
  parameters:
    FamilyId:
      name: familyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    StudyId:
      name: studyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Appointment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        scheduledFor:
          type: string
          format: date-time
        status:
          type: string
    AppointmentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
    AppointmentCreateInput:
      type: object
      required: [scheduledFor]
      properties:
        familyId:
          type: string
          format: uuid
        memberId:
          type: string
          format: uuid
          nullable: true
        scheduledFor:
          type: string
          format: date-time
...
    Study:
      type: object
      properties:
        id:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        type:
          type: string
        status:
          type: string
        name:
          type: string
    StudyListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Study"
    StudyCreateInput:
      type: object
      required: [type, name]
      properties:
        familyId:
          type: string
          format: uuid
        memberId:
          type: string
          format: uuid
          nullable: true
        type:
          type: string
        status:
          type: string
        name:
          type: string
        requestedAt:
          type: string
          format: date-time
          nullable: true
    StudyUpdateInput:
      allOf:
        - $ref: "#/components/schemas/StudyCreateInput"
...